+++ src/core/config/config.service.ts
@@ -81,6 +81,7 @@ export class ConfigService {
     host?: '::' | '0.0.0.0' | string
     auth: 'form' | 'none'
     theme: string
+    lightingMode: 'auto' | 'light' | 'dark'
     sudo?: boolean
     restart?: string
     lang?: string
@@ -214,7 +215,8 @@ export class ConfigService {
         recommendChildBridges: this.recommendChildBridges,
       },
       formAuth: Boolean(this.ui.auth !== 'none'),
-      theme: this.ui.theme || 'auto',
+      theme: this.ui.theme || 'orange',
+      lightingMode: this.ui.lightingMode || 'auto',
       serverTimestamp: new Date().toISOString(),
     }
   }